orgnr som string istedenfor int
