+++ frontend/mr-admin-flate/src/components/avtaler/opprett/OpprettAvtaleContainer.module.scss
@@ -33,12 +33,3 @@
     grid-template-columns: repeat(2, 1fr);
   }
 }
-
-.icon_text_align {
-  display: flex;
-  align-items: center;
-}
-
-.virksomhet {
-  margin-top: 0.5rem;
-}
