+++ frontend/mr-admin-flate/src/components/avtaler/opprett/OpprettAvtaleContainer.tsx
@@ -1,5 +1,4 @@
 import { zodResolver } from "@hookform/resolvers/zod";
-import { CheckmarkIcon } from "@navikt/aksel-icons";
 import { Button, Select, TextField } from "@navikt/ds-react";
 import classNames from "classnames";
 import { Avtale, AvtaleRequest, Avtaletype } from "mulighetsrommet-api-client";
@@ -8,27 +7,15 @@ import { NavEnhet } from "mulighetsrommet-api-client/build/models/NavEnhet";
 import { Tiltakstype } from "mulighetsrommet-api-client/build/models/Tiltakstype";
 import { StatusModal } from "mulighetsrommet-veileder-flate/src/components/modal/delemodal/StatusModal";
 import { porten } from "mulighetsrommet-veileder-flate/src/constants";
-import {
-  Dispatch,
-  ReactNode,
-  SetStateAction,
-  useEffect,
-  useReducer,
-  useState,
-} from "react";
+import { Dispatch, ReactNode, SetStateAction, useState } from "react";
 import { FormProvider, SubmitHandler, useForm } from "react-hook-form";
 import z from "zod";
 import { mulighetsrommetClient } from "../../../api/clients";
 import { useNavigerTilAvtale } from "../../../hooks/useNavigerTilAvtale";
-import {
-  capitalize,
-  capitalizeEveryWord,
-  formaterDatoSomYYYYMMDD,
-} from "../../../utils/Utils";
-import { Laster } from "../../laster/Laster";
+import { capitalize, formaterDatoSomYYYYMMDD } from "../../../utils/Utils";
 import { Datovelger } from "../../skjema/OpprettComponents";
+import { VirksomhetInput } from "../../virksomhet/VirksomhetInput";
 import styles from "./OpprettAvtaleContainer.module.scss";
-import { initialState, reducer } from "./virksomhetReducer";
 
 interface OpprettAvtaleContainerProps {
   onAvbryt: () => void;
@@ -87,19 +74,9 @@ export function OpprettAvtaleContainer({
   const redigeringsModus = !!avtale;
   const [feil, setFeil] = useState<string | null>("");
 
-  const [virksomhetState, virksomhetDispatcher] = useReducer(
-    reducer,
-    initialState
-  );
   const clickCancel = () => {
     setFeil(null);
   };
-
-  useEffect(() => {
-    if (avtale?.leverandor?.organisasjonsnummer?.length === 9) {
-      sjekkOrgnr(avtale.leverandor.organisasjonsnummer);
-    }
-  }, [avtale?.leverandor]);
 
   const form = useForm<inferredSchema>({
     resolver: zodResolver(Schema),
@@ -120,8 +97,6 @@ export function OpprettAvtaleContainer({
     register,
     handleSubmit,
     formState: { errors },
-    setError,
-    clearErrors,
   } = form;
 
   const postData: SubmitHandler<inferredSchema> = async (
@@ -156,27 +131,6 @@ export function OpprettAvtaleContainer({
       return;
     } catch {
       setFeil("Klarte ikke opprette eller redigere avtale");
-    }
-  };
-
-  const sjekkOrgnr = async (orgnr: string) => {
-    virksomhetDispatcher({ type: "Reset" });
-    if (orgnr.trim().length === 9) {
-      virksomhetDispatcher({ type: "Hent data" });
-      try {
-        const response =
-          await mulighetsrommetClient.hentVirksomhet.hentVirksomhetMedOrgnr({
-            orgnr: orgnr.trim(),
-          });
-
-        virksomhetDispatcher({ type: "Data hentet", payload: response });
-        clearErrors("leverandor");
-      } catch (error) {
-        virksomhetDispatcher({ type: "Reset" });
-        setError("leverandor", {
-          message: `Fant ikke leverandør med nummer: ${orgnr}`,
-        });
-      }
     }
   };
 
@@ -258,30 +212,8 @@ export function OpprettAvtaleContainer({
             label="Antall plasser"
             {...register("antallPlasser", { valueAsNumber: true })}
           />
-          <div>
-            <TextField
-              error={errors.leverandor?.message}
-              label={"Leverandør"}
-              {...register("leverandor", {
-                onChange: (e: React.ChangeEvent<HTMLInputElement>) =>
-                  sjekkOrgnr(e.currentTarget.value),
-              })}
-            />
-            <div className={styles.virksomhet}>
-              {virksomhetState.status === "fetching" ? (
-                <Laster tekst="Henter virksomhet" sentrert={false} />
-              ) : null}
-              {virksomhetState.status === "fetched" ? (
-                <span
-                  className={styles.icon_text_align}
-                  aria-label={`Fant virksomhet for orgnr: ${virksomhetState.data?.organisasjonsnummer} med navn ${virksomhetState.data?.navn}`}
-                >
-                  <CheckmarkIcon color="green" title="Fant virksomhet" />{" "}
-                  {capitalizeEveryWord(virksomhetState.data?.navn)}
-                </span>
-              ) : null}
-            </div>
-          </div>
+
+          <VirksomhetInput avtale={avtale} />
           <Select
             error={errors.avtaletype?.message}
             label={"Avtaletype"}
