+++ frontend/mr-admin-flate/src/components/virksomhet/VirksomhetInput.module.scss
@@ -1,0 +1,8 @@
+.icon_text_align {
+  display: flex;
+  align-items: center;
+}
+
+.virksomhet {
+  margin-top: 0.5rem;
+}
