+++ frontend/mr-admin-flate/src/mocks/fixtures/mock_virksomhet.ts
@@ -1,0 +1,8 @@
+import { Virksomhet } from "mulighetsrommet-api-client";
+
+export const mockVirksomhet: Virksomhet = {
+  navn: "Joblearn AS - Avd. Bergen",
+  organisasjonsnummer: "876456342",
+  overordnetEnhetNavn: "Joblearn AS",
+  overordnetEnhetOrganisasjonsnummer: "876456344",
+};
