+++ mulighetsrommet-api/src/main/kotlin/no/nav/mulighetsrommet/api/clients/enhetsregister/AmtEnhetsregisterClient.kt
@@ -1,5 +1,5 @@
 package no.nav.mulighetsrommet.api.clients.enhetsregister
 
 interface AmtEnhetsregisterClient {
-    suspend fun hentVirksomhet(virksomhetsnummer: Int): VirksomhetDto?
+    suspend fun hentVirksomhet(virksomhetsnummer: String): VirksomhetDto?
 }
