+++ mulighetsrommet-api/src/test/kotlin/no/nav/mulighetsrommet/api/services/ArrangorServiceTest.kt
@@ -15,13 +15,13 @@ class ArrangorServiceTest : FunSpec({
     val arrangorService = ArrangorService(amtEnhetsregister)
 
     beforeSpec {
-        coEvery { amtEnhetsregister.hentVirksomhet(111) } returns VirksomhetDto(
+        coEvery { amtEnhetsregister.hentVirksomhet("111") } returns VirksomhetDto(
             organisasjonsnummer = "789",
             navn = "Bedrift 1",
             overordnetEnhetOrganisasjonsnummer = "1011",
             overordnetEnhetNavn = "Overordnetbedrift 1"
         )
-        coEvery { amtEnhetsregister.hentVirksomhet(222) } returns VirksomhetDto(
+        coEvery { amtEnhetsregister.hentVirksomhet("222") } returns VirksomhetDto(
             organisasjonsnummer = "7891",
             navn = "Bedrift 2",
             overordnetEnhetOrganisasjonsnummer = "1011",
